requirements.txt
strawberry-graphql==0.248.1
